+++ python/lsst/ts/ess/controller/__init__.py
@@ -35,5 +35,3 @@ from .command_handler import *
 
 # Import sub modules
 from . import device
-from . import schema
-from . import sensor