+++ python/lsst/ts/ess/controller/device/vcp_ftdi.py
@@ -23,24 +23,23 @@ __all__ = ["VcpFtdi"]
 
 import asyncio
 import logging
-from typing import Callable, Tuple
+from typing import Callable
 
-from pylibftdi import Device, FtdiError
+from pylibftdi import Device
 
-from .base_device import BaseDevice
-from ..sensor import BaseSensor
+from lsst.ts.ess import common
 
 
-class VcpFtdi(BaseDevice):
+class VcpFtdi(common.device.BaseDevice):
     """USB Virtual Communications Port (VCP) for FTDI device.
 
     Parameters
     ----------
-    device_id: `str`
+    device_id : `str`
         The hardware device ID to connect to. This can be a physical ID (e.g.
         /dev/ttyUSB0), a serial port (e.g. serial_ch_1) or any other ID used by
         the specific device.
-    sensor: `BaseSensor`
+    sensor : `common.sensor.BaseSensor`
         The sensor that produces the telemetry.
     callback_func : `Callable`
         Callback function to receive instrument output.
@@ -50,7 +49,7 @@ class VcpFtdi(BaseDevice):
         self,
         name: str,
         device_id: str,
-        sensor: BaseSensor,
+        sensor: common.sensor.BaseSensor,
         callback_func: Callable,
         log: logging.Logger,
     ) -> None: