+++ doc/version_history.rst
@@ -10,9 +10,17 @@ v0.4.0
 ======
 
 * Replaced the use of ts_salobj functions with ts_utils functions.
+* Moved all device reply validating code to ts.ess.common.
+* Moved all sensors code from ts.ess.controller to ts.ess.common.
+* Moved code to determine what sensor is connected from ts.ess.controller to ts.ess.common.
+* Moved BaseDevice and MockDevice from ts.ess.controller to ts.ess.common.
+* Removed all obsolete schema related code since it also is in ts.ess.common.
+* Updated the documentation to reflect all sensor and device code changes.
+* Moved most of the command handler code and the socket server unit test from ts.ess.controller to ts.ess.common.
 
 Requires:
 
+* ts_ess_common
 * ts_tcpip 0.3
 * ts_utils 1.0
 
@@ -29,6 +37,7 @@ v0.3.0
 
 Requires:
 
+* ts_ess_common
 * ts_tcpip 0.3.1